requirements.txt
streamlit
pandas
sqlalchemy
bcrypt
cryptography
pyotp
python-dotenv
